requirements.txt
flask>=2.3.0
flask-sqlalchemy>=3.0.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
gunicorn>=21.0.0
sqlalchemy>=2.0.0
requests>=2.31.0
schedule>=1.2.0
pillow>=10.0.0
qrcode>=7.4.0
aiohttp>=3.8.0
trafilatura>=1.6.0
psutil>=5.9.0
alembic>=1.12.0
email-validator>=2.0.0
werkzeug>=2.3.0
python-dotenv
telegram
python-dotenv
python-dotenv
